An initial basic mock user service with OIDC compliant flows
Makes heavy use of the [node-oidc-provider](https://github.com/panva/node-oidc-provider) library, which provides dev interactions/flows out of the box (fully OIDC compliant). Any username and password combination is permitted for logging in, making this very useful for local development and CI. In the future, we could extend (and make more configurable) this mock user service to support different scenarios for testing etc.

Note: included is a patch to the node-oidc-provider which relaxes a particular constraint (see the file in the `patches` folder for details) – this is purely to unblock usage of the [keycloak-gatekeeper](https://github.com/keycloak/keycloak-gatekeeper).